feat: Add speed selection and reset grid functionality, and RunVisualizer function to Nav component
This commit adds the ability to select the speed of the visualization and reset the grid in the Nav component. It introduces new state variables for speed and isDisabled, as well as new functions handleResetGrid and handlerRunVisualizer. These additions enhance the user experience by allowing them to control the speed of the visualization and easily reset the grid when needed.